+++ CHANGELOG.md
@@ -2,5 +2,9 @@
 # Change Log
 
 
+## [1.0.0]
+- Improved list display to use GridField (not editable)
+- Stable build
+
 ## [0.0.1]
 - Initial commit